Added masked SELECT actions for all tables in database
Upon finding a table, the RL agent will be able to run a SELECT * FROM {table_name} query, where {table_name} is the table's name.

If the table has not yet been found, the agent receives a negative reward and the query is not run. The current state is returned instead.